+++ src/main/java/com/unascribed/yttr/mixin/client/MixinMinecraftClient.java
@@ -6,13 +6,16 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
+import com.unascribed.yttr.compat.modmenu.YttrConfigScreen;
 import com.unascribed.yttr.network.MessageC2SAttack;
 import com.unascribed.yttr.util.Attackable;
 
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.MinecraftClient;
+import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.network.ClientPlayerEntity;
+import net.minecraft.sound.MusicSound;
 
 @Environment(EnvType.CLIENT)
 @Mixin(MinecraftClient.class)
@@ -20,6 +23,8 @@ public class MixinMinecraftClient {
 
 	@Shadow
 	public ClientPlayerEntity player;
+	@Shadow
+	public Screen currentScreen;
 	
 	@Inject(at=@At("HEAD"), method="doAttack")
 	private void doAttack(CallbackInfoReturnable<Boolean> ci) {
@@ -27,4 +32,12 @@ public class MixinMinecraftClient {
 			new MessageC2SAttack().sendToServer();
 		}
 	}
+	
+	@Inject(at=@At("HEAD"), method="getMusicType", cancellable=true)
+	public void getMusicType(CallbackInfoReturnable<MusicSound> ci) {
+		if (currentScreen instanceof YttrConfigScreen) {
+			ci.setReturnValue(YttrConfigScreen.MUSIC);
+		}
+	}
+	
 }