+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 * Now works on servers, oops
 * Fixed conditional items not showing in creative tabs
 * Fixed Squeeze Trees not generating
+* Added Scorched retrogen
+* Added a config screen (incomplete)